+++ src/main/java/security/validators/FlaskDebugModeInspection.kt
@@ -7,8 +7,10 @@ import com.jetbrains.python.inspections.PyInspection
 import com.jetbrains.python.inspections.PyInspectionVisitor
 import com.jetbrains.python.psi.PyBoolLiteralExpression
 import com.jetbrains.python.psi.PyCallExpression
+import com.jetbrains.python.psi.PyFile
 import com.jetbrains.python.psi.PyReferenceExpression
 import security.Checks
+import security.helpers.ImportValidators.hasImportedNamespace
 
 class FlaskDebugModeInspection : PyInspection() {
     val check = Checks.FlaskDebugModeCheck;
@@ -25,6 +27,8 @@ class FlaskDebugModeInspection : PyInspection() {
         override fun visitPyCallExpression(node: PyCallExpression) {
             val calleeName = node.callee?.name ?: return
             if (calleeName != "run") return
+            if (node.containingFile !is PyFile) return
+            if (!hasImportedNamespace(node.containingFile as PyFile, "flask")) return
             if (node.firstChild !is PyReferenceExpression) return
             if ((node.firstChild as PyReferenceExpression).asQualifiedName().toString() != "app.run") return
             val debugArg = node.getKeywordArgument("debug") ?: return