+++ src/test/java/security/validators/DjangoRawSqlInspectionTest.kt
@@ -51,4 +51,58 @@ class DjangoRawSqlInspectionTest: SecurityTestTask() {
         """.trimIndent()
         testCodeCallExpression(code, 0, Checks.DjangoRawSqlCheck, "test.py", DjangoRawSqlInspection())
     }
+
+    @Test
+    fun `test cursor execute with no quotes`(){
+        var code = """
+            from django.db import connection
+
+            def my_custom_sql(self):
+                with connection.cursor() as cursor:
+                    cursor.execute("UPDATE bar SET foo = 1 WHERE baz = %s", [self.baz])
+                    cursor.execute("SELECT foo FROM bar WHERE baz = %s", [self.baz])
+                    row = cursor.fetchone()
+                return row
+        """.trimIndent()
+        testCodeCallExpression(code, 0, Checks.DjangoRawSqlCheck, "test.py", DjangoRawSqlInspection())
+    }
+
+    @Test
+    fun `test cursor execute with quotes`(){
+        var code = """
+            from django.db import connection
+
+            def my_custom_sql(self):
+                with connection.cursor() as cursor:
+                    cursor.execute("UPDATE bar SET foo = 1 WHERE baz = %s", [self.baz])
+                    cursor.execute("SELECT foo FROM bar WHERE baz = '%s'", [self.baz])
+                    row = cursor.fetchone()
+                return row
+        """.trimIndent()
+        testCodeCallExpression(code, 1, Checks.DjangoRawSqlCheck, "test.py", DjangoRawSqlInspection())
+    }
+
+    @Test
+    fun `test model raw with quotes`(){
+        var code = """
+            from django.db import connection
+            from .models import User
+
+            def my_view(self):
+                User.objects.raw("SELECT * FROM myapp_person WHERE last_name = '%s'", [lname])
+        """.trimIndent()
+        testCodeCallExpression(code, 1, Checks.DjangoRawSqlCheck, "test.py", DjangoRawSqlInspection())
+    }
+
+    @Test
+    fun `test model raw with no quotes`(){
+        var code = """
+            from django.db import connection
+            from .models import User
+
+            def my_view(self):
+                User.objects.raw("SELECT * FROM myapp_person WHERE last_name = %s", [lname])
+        """.trimIndent()
+        testCodeCallExpression(code, 0, Checks.DjangoRawSqlCheck, "test.py", DjangoRawSqlInspection())
+    }
 }